+++ Sample/src/minimum/java/com.live2d.demo.minimum/LAppMinimumSprite.java
@@ -55,8 +55,6 @@ public class LAppMinimumSprite {
      * @param uvVertex uv頂点座標
      */
     public void renderImmediate(int textureId, final float[] uvVertex) {
-        GLES20.glEnable(GL_TEXTURE_2D);
-
         // attribute属性を有効にする
         GLES20.glEnableVertexAttribArray(positionLocation);
         GLES20.glEnableVertexAttribArray(uvLocation);
@@ -91,7 +89,7 @@ public class LAppMinimumSprite {
     public void resize(float x, float y, float width, float height) {
         rect.left = x - width * 0.5f;
         rect.right = x + width * 0.5f;
-        rect.up = x + height * 0.5f;
+        rect.up = y + height * 0.5f;
         rect.down = y - height * 0.5f;
     }