+++ CHANGELOG.md
@@ -4,8 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [4-r.1-alpha.1] - 2022-10-06
+## [4-r.1-beta.1] - 2022-12-08
+
+### Fixed
+
+* Fix `resize` function in `LAppSprite` because resizing calculation is wrong.
+* Remove the glEnable(GL_TEXTURE_2D) instruction, which was not valid.
+* Fix `onUpdate` function in `LAppLive2DManager` not to create an instance of CubismMatrix44.
+
+## 4-r.1-alpha.1 - 2022-10-06
 
 ### Added
 
 * New released!
+
+[4-r.1-beta.1]: https://github.com/Live2D/CubismJavaSamples/compare/4-r.1-alpha.1...4-r.1-beta.1